support se tests on functions
